requirements.txt
chromadb>=1.3.0
ollama>=0.6.1
python-dotenv>=1.0.0
tqdm>=4.60.0
nltk>=3.8.1
